Fix sonar check and others improvements
Created new class ItemReadMapAttributes (derived from Item)
To break the function, which was very large, in smaller functions, improving the reading and organization of the file item.cpp

Converted a lot of "iterator" pointer implicit conversion to "std::to_address"
Moved a lot of functions from ".h" to ".cpp", to avoid that when modifying an include file we have to recompile a large number of files
Fixed various init-statement sonar checks
Converted a lot of "find" and similiar map functions to "contains"
Converted a lot of "std::find" to"std::ranges::find"
Converted a lot of "std::any_of" to"std::ranges::any_of"
Converted a lot of "std::sort" to"std::ranges::sort"
Converted a lot of "std::for_each" to"std::ranges::for_each"
Converted a lot of "std::shuffle" to"std::ranges::shuffle"
Converted a lot of "std::count" to"std::ranges::count"
Converted a lot of "std::find_if" to"std::ranges::find_if"

Added default operator <=> in position.h instead of manual verifications
Renamed various variables and added structured binding for some std::map loops (iobestiary.cpp, protocolgame.cpp)

Indentation of various functions

+++ src/items/bed.cpp
@@ -30,7 +30,7 @@ void BedItem::serializeAttr(PropWriteStream& propWriteStream) const
 	if (getSleepStart() != 0) {
 		propWriteStream.write<uint8_t>(ATTR_SLEEPSTART);
 		// FIXME: should be stored as 64-bit, but we need to retain backwards compatibility
-		propWriteStream.write<uint32_t>(static_cast<uint32_t>(getSleepStart()));
+		propWriteStream.write<uint32_t>(getSleepStart());
 	}
 }
